+++ src/euler.hpp
@@ -35,57 +35,57 @@
 
 namespace euler {
 
-void evalSplitting(const double *conservativeL, const double *conservativeR, const double *n, double *fluxes, double *lambda);
+void evalSplitting(const float *conservativeL, const float *conservativeR, const float *n, float *fluxes, float *lambda);
 
-void evalFluxes(const double *conservative, const double *primitive, const double *n, double *fluxes);
+void evalFluxes(const float *conservative, const float *primitive, const float *n, float *fluxes);
 
 void computeRHS(problem::ProblemType problemType, ComputationInfo &computationInfo,
                 const int order, const ScalarStorage<int> &solvedBoundaryInterfaceBCs,
-                const ScalarPiercedStorage<double> &cellConservatives, ScalarPiercedStorage<double> *cellsRHS, double *maxEig);
+                const ScalarPiercedStorage<float> &cellConservatives, ScalarPiercedStorage<float> *cellsRHS, float *maxEig);
 
 #if ENABLE_CUDA
 void cuda_initialize();
 void cuda_finalize();
 #endif
 
-void resetRHS(ScalarPiercedStorage<double> *cellsRHS);
+void resetRHS(ScalarPiercedStorage<float> *cellsRHS);
 #if ENABLE_CUDA
-void cuda_resetRHS(ScalarPiercedStorage<double> *cellsRHS);
+void cuda_resetRHS(ScalarPiercedStorage<float> *cellsRHS);
 #endif
 
 void updateRHS(problem::ProblemType problemType, ComputationInfo &computationInfo,
                const int order, const ScalarStorage<int> &solvedBoundaryInterfaceBCs,
-               const ScalarPiercedStorage<double> &cellConservatives, ScalarPiercedStorage<double> *cellsRHS, double *maxEig);
+               const ScalarPiercedStorage<float> &cellConservatives, ScalarPiercedStorage<float> *cellsRHS, float *maxEig);
 #if ENABLE_CUDA
 void cuda_updateRHS(problem::ProblemType problemType, ComputationInfo &computationInfo,
                     const int order, const ScalarStorage<int> &solvedBoundaryInterfaceBCs,
-                    const ScalarPiercedStorage<double> &cellConservatives, ScalarPiercedStorage<double> *cellsRHS, double *maxEig);
+                    const ScalarPiercedStorage<float> &cellConservatives, ScalarPiercedStorage<float> *cellsRHS, float *maxEig);
 #endif
 
 void evalInterfaceBCValues(problem::ProblemType problemType, int BCType,
-                           const std::array<double, 3> &point,
-                           const std::array<double, 3> &normal,
-                           const double *conservative, double *conservative_BC);
+                           const std::array<float, 3> &point,
+                           const std::array<float, 3> &normal,
+                           const float *conservative, float *conservative_BC);
 
-void evalFreeFlowBCValues(const std::array<double, 3> &point,
-                          const std::array<double, 3> &normal,
-                          const std::array<double, BC_INFO_SIZE> &info,
-                          const double *conservative, double *conservative_BC);
+void evalFreeFlowBCValues(const std::array<float, 3> &point,
+                          const std::array<float, 3> &normal,
+                          const std::array<float, BC_INFO_SIZE> &info,
+                          const float *conservative, float *conservative_BC);
 
-void evalReflectingBCValues(const std::array<double, 3> &point,
-                          const std::array<double, 3> &normal,
-                          const std::array<double, BC_INFO_SIZE> &info,
-                          const double *conservative, double *conservative_BC);
+void evalReflectingBCValues(const std::array<float, 3> &point,
+                          const std::array<float, 3> &normal,
+                          const std::array<float, BC_INFO_SIZE> &info,
+                          const float *conservative, float *conservative_BC);
 
-void evalWallBCValues(const std::array<double, 3> &point,
-                          const std::array<double, 3> &normal,
-                          const std::array<double, BC_INFO_SIZE> &info,
-                          const double *conservative, double *conservative_BC);
+void evalWallBCValues(const std::array<float, 3> &point,
+                          const std::array<float, 3> &normal,
+                          const std::array<float, BC_INFO_SIZE> &info,
+                          const float *conservative, float *conservative_BC);
 
-void evalDirichletBCValues(const std::array<double, 3> &point,
-                          const std::array<double, 3> &normal,
-                          const std::array<double, BC_INFO_SIZE> &info,
-                          const double *conservative, double *conservative_BC);
+void evalDirichletBCValues(const std::array<float, 3> &point,
+                          const std::array<float, 3> &normal,
+                          const std::array<float, BC_INFO_SIZE> &info,
+                          const float *conservative, float *conservative_BC);
 
 }
